SceneIdeas: Convert all element array handling from unsigned int to unsigned short. OpenGL ES 2.0 only allows GL_UNSIGNED_BYTE or GL_UNSIGNED_SHORT for the index type, and the use of GL_UNSIGNED_INT, while legal for OpenGL and permitted by several OpenGL ES 2.0 implmentations generates an error on stricter ones.

+++ src/scene-ideas/a.cc
@@ -170,10 +170,10 @@ LetterA::LetterA()
     // Primitive state so that the draw call can issue the primitives we want.
     unsigned int curOffset(0);
     primVec_.push_back(PrimitiveState(GL_TRIANGLE_STRIP, 28, curOffset));
-    curOffset += (28 * sizeof(unsigned int));
+    curOffset += (28 * sizeof(unsigned short));
     primVec_.push_back(PrimitiveState(GL_TRIANGLE_STRIP, 17, curOffset));
-    curOffset += (17 * sizeof(unsigned int));
+    curOffset += (17 * sizeof(unsigned short));
     primVec_.push_back(PrimitiveState(GL_LINE_STRIP, 28, curOffset));
-    curOffset += (28 * sizeof(unsigned int));
+    curOffset += (28 * sizeof(unsigned short));
     primVec_.push_back(PrimitiveState(GL_LINE_STRIP, 17, curOffset));
 }
