+++ src/scene-ideas/s.cc
@@ -125,6 +125,6 @@ LetterS::LetterS()
     // Primitive state so that the draw call can issue the primitives we want.
     unsigned int curOffset(0);
     primVec_.push_back(PrimitiveState(GL_TRIANGLE_STRIP, 30, curOffset));
-    curOffset += (30 * sizeof(unsigned int));
+    curOffset += (30 * sizeof(unsigned short));
     primVec_.push_back(PrimitiveState(GL_LINE_STRIP, 30, curOffset));
 }
